feat: Use iced 0.14

+++ toboggan-desktop/src/widgets.rs
@@ -8,7 +8,7 @@ use crate::constants::{
 use crate::message::Message;
 
 // Helper function to create text with consistent styling
-pub fn create_text(content: &str, size: u16) -> Text<'_> {
+pub fn create_text(content: &str, size: f32) -> Text<'_> {
     text(content).size(size)
 }
